Refactor Hauki export related functions to ReservationUnitHaukiExporter
